Add files via upload
You have been hired by an IT company as an IT subject matter expert. Management has a client who
wants to upgrade their IT systems, but the client has a limited budget and no more physical space to
install new equipment. You suggest implementing virtual machines instead of more hardware to help
reduce costs and to meet the client’s operational needs. Part of the upgrade includes installing a
Firewall, a PC for the CEO, a Webserver, DNS server, and some Linux machines for monitoring and
conducting cybersecurity related tasks. Once the new network is installed, you will document
configurations and create a security baseline by scanning for open port